fix(gui): move automatic worktree cleanup into a toolbar chip
The Automatic cleanup card sat above the worktree inventory in a
fixed-height panel, so every row it spent unconditionally was a worktree
the list below could not show - for a policy that is consulted rarely and
changed almost never. Replace it with one chip in the inventory toolbar's
leading slot, opening a popover that carries the switch, the safety
sentence, the threshold, the schedule and the History link. Drop the
panel header's subtitle with it, so the list card is the only child of
the fill-height column in both views.

The chip states the policy at a glance (Cleanup - On - 7d with a green
dot, Cleanup off with a grey one, a bare Cleanup with no dot until the
read lands), so the policy read stays eager. The gate ladder is
unchanged: absent renders nothing, and checking/offline/unsupported
render the chip inert with the same sentence in a tooltip and no
popover - aria-disabled rather than disabled, because a disabled button
takes neither pointer events nor focus and that sentence is the whole
point of the chip in those states.

The write path moves verbatim: the revision guard, the bounds-driven
validation, the conflict re-read, and the loading/error lines. The Sweep
dialog's host-scoped focus request is retargeted rather than removed - a
matching host now OPENS the popover and moves the caret onto the switch,
then clears the request.
