--- clients/gui-app/src/components/settings/panels/__tests__/worktree-auto-cleanup-section.test.tsx
@@ -1,765 +1,0 @@
-import {
-  afterEach,
-  beforeEach,
-  describe,
-  expect,
-  it,
-  vi,
-  type MockInstance,
-} from "vitest";
-import {
-  cleanup,
-  fireEvent,
-  render,
-  screen,
-  waitFor,
-  type RenderResult,
-} from "@testing-library/react";
-import type { ReactNode } from "react";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { HostClient } from "@traycer-clients/shared/host-client/host-client";
-import { mockLocalHostEntry } from "@traycer-clients/shared/host-client/mock/mock-host-directory";
-import { MockHostMessenger } from "@traycer-clients/shared/host-client/mock/mock-host-messenger";
-import { createRequestContextFixture } from "@traycer-clients/shared/test-fixtures/request-context";
-import { HostRpcError } from "@traycer-clients/shared/host-transport/host-messenger";
-import type { WorktreeAutoCleanupPolicyState } from "@traycer/protocol/host/worktree-auto-cleanup-schemas";
-import { hostRpcRegistry, type HostRpcRegistry } from "@/lib/host";
-import { TooltipProvider } from "@/components/ui/tooltip";
-
-/**
- * The automatic-cleanup card's state ladder and its one write.
- *
- * The RPCs run for real against a `MockHostMessenger`, so the revision guard,
- * the bounds-driven validation and the conflict re-read all go through the
- * actual query/mutation path rather than a hook stub. Only the two per-host
- * FACTS the card cannot synthesize — reachability and whether this host
- * advertised the capability — are mocked.
- */
-const state = vi.hoisted(() => ({
-  reachability: { status: "reachable", hostLabel: "Host A" },
-  supported: true as boolean | null,
-}));
-
-vi.mock("@/hooks/agent/use-host-reachability", () => ({
-  useHostReachability: () => state.reachability,
-}));
-
-vi.mock("@/hooks/host/use-host-supports-method", () => ({
-  useHostMethodSupport: () => state.supported,
-  useHostSupportsMethod: () => state.supported === true,
-  useHostMethodSchemaVersion: () => null,
-}));
-
-import { WorktreeAutoCleanupSection } from "@/components/settings/panels/worktree-auto-cleanup-section";
-import {
-  hostScopeFixture,
-  hostScopeOptionFixture,
-} from "@/components/settings/host-scope/host-scope-fixture";
-import { useWorktreeCleanupViewStore } from "@/stores/settings/worktree-cleanup-view-store";
-
-/**
- * The toggle renders immediately but stays DISABLED until the policy read
- * lands - the control must never write a revision it has not read. Tests that
- * press it therefore wait for the loaded state, not merely for the element.
- */
-async function loadedCleanupToggle(): Promise<HTMLElement> {
-  await waitFor(() => {
-    expect(
-      screen
-        .getByRole("switch", { name: "Automatic cleanup" })
-        .hasAttribute("disabled"),
-    ).toBe(false);
-  });
-  return screen.getByRole("switch", { name: "Automatic cleanup" });
-}
-
-/**
- * The threshold controls live behind a default-collapsed disclosure, so every
- * test that touches them has to open it first - which is also the assertion
- * that the trigger is the only way in.
- */
-async function openThresholdEditor(): Promise<void> {
-  const trigger = await screen.findByRole("button", {
-    name: "Configure automatic cleanup",
-  });
-  fireEvent.click(trigger);
-  await waitFor(() => {
-    screen.getByRole("textbox", { name: "Custom inactivity days" });
-  });
-}
-
-/**
- * Which nodes were scrolled to. The jsdom setup installs a no-op
- * `scrollIntoView` on the prototype; spying on it records the receiver in
- * `mock.contexts`, which is the only thing these tests need from it.
- */
-function spyOnScrollIntoView(): MockInstance<() => void> {
-  return vi.spyOn(Element.prototype, "scrollIntoView");
-}
-
-function policyFixture(
-  overrides: Partial<WorktreeAutoCleanupPolicyState>,
-): WorktreeAutoCleanupPolicyState {
-  return {
-    enabled: false,
-    inactivityDays: 30,
-    revision: 0,
-    updatedAt: null,
-    updatedByUserId: null,
-    lastEvaluatedAt: null,
-    nextEvaluationAt: null,
-    pausedReason: null,
-    bounds: { minDays: 1, maxDays: 365 },
-    ...overrides,
-  };
-}
-
-interface PolicyHandlers {
-  readonly get: () => WorktreeAutoCleanupPolicyState;
-  readonly set: (request: {
-    readonly enabled: boolean;
-    readonly inactivityDays: number;
-    readonly expectedRevision: number;
-  }) => WorktreeAutoCleanupPolicyState;
-}
-
-function clientWithPolicy(
-  handlers: PolicyHandlers,
-): HostClient<HostRpcRegistry> {
-  const spine = new HostClient<HostRpcRegistry>({
-    registry: hostRpcRegistry,
-    invalidator: { invalidateHostScope: () => undefined },
-    findHostById: (hostId) =>
-      hostId === mockLocalHostEntry.hostId ? mockLocalHostEntry : null,
-    messenger: new MockHostMessenger<HostRpcRegistry>({
-      registry: hostRpcRegistry,
-      requestId: () => `req-${Math.random().toString(36).slice(2)}`,
-      handlers: {
-        "worktree.getAutoCleanupPolicy": () => handlers.get(),
-        "worktree.setAutoCleanupPolicy": (request) => handlers.set(request),
-      },
-    }),
-  });
-  spine.setRequestContext(
-    createRequestContextFixture({ origin: "renderer", bearerToken: "tok-1" }),
-  );
-  return spine.createRequester(mockLocalHostEntry);
-}
-
-function sectionForHost(
-  client: HostClient<HostRpcRegistry> | null,
-  host: { readonly hostId: string; readonly name: string },
-): ReactNode {
-  return (
-    <WorktreeAutoCleanupSection
-      scope={hostScopeFixture({
-        host: hostScopeOptionFixture(host),
-        client,
-      })}
-      onOpenHistory={() => undefined}
-    />
-  );
-}
-
-const HOST_A = { hostId: "host-a", name: "Host A" };
-const HOST_B = { hostId: "host-b", name: "Host B" };
-
-function renderSection(
-  client: HostClient<HostRpcRegistry> | null,
-): RenderResult {
-  const queryClient = new QueryClient({
-    defaultOptions: { queries: { retry: false }, mutations: { retry: false } },
-  });
-  const Wrapper = (props: { readonly children: ReactNode }): ReactNode => (
-    <QueryClientProvider client={queryClient}>
-      <TooltipProvider>{props.children}</TooltipProvider>
-    </QueryClientProvider>
-  );
-  return render(sectionForHost(client, HOST_A), { wrapper: Wrapper });
-}
-
-beforeEach(() => {
-  state.reachability = { status: "reachable", hostLabel: "Host A" };
-  state.supported = true;
-  useWorktreeCleanupViewStore.setState({
-    view: "settings",
-    focusedRunId: null,
-    autoCleanupFocusHostId: null,
-  });
-});
-
-afterEach(() => {
-  cleanup();
-  vi.restoreAllMocks();
-});
-
-describe("WorktreeAutoCleanupSection", () => {
-  it("renders default-off controls on a supported host", async () => {
-    renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({}),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    const toggle = await loadedCleanupToggle();
-    expect(toggle.getAttribute("data-state")).toBe("unchecked");
-    // Default-off says so in words, and the threshold control is not offered
-    // for a policy that deletes nothing.
-    expect(
-      screen.getByTestId("worktree-auto-cleanup-summary").textContent,
-    ).toBe("Automatic cleanup · Off");
-    screen.getByText("Nothing is deleted automatically.");
-    expect(
-      screen.queryByRole("textbox", { name: "Custom inactivity days" }),
-    ).toBeNull();
-  });
-
-  it("offers no way to configure a policy that is switched off", async () => {
-    renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({}),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    await loadedCleanupToggle();
-    // Nothing to unfold while cleanup deletes nothing: a disclosure over an
-    // inert threshold is a control that cannot change an outcome.
-    expect(screen.queryByRole("button", { name: /configure/i })).toBeNull();
-    expect(screen.queryByRole("button", { name: /7 days/ })).toBeNull();
-  });
-
-  it("starts collapsed for an enabled policy and keeps the switch reachable", async () => {
-    renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({ enabled: true, inactivityDays: 30 }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    const toggle = await loadedCleanupToggle();
-    expect(toggle.getAttribute("data-state")).toBe("checked");
-    // The summary carries the threshold, so the collapsed card still answers
-    // "after how long" without unfolding anything.
-    expect(
-      screen.getByTestId("worktree-auto-cleanup-summary").textContent,
-    ).toBe("Automatic cleanup · On · after 30 days");
-    for (const days of [7, 14, 30, 60, 90]) {
-      expect(screen.queryByRole("button", { name: `${days} days` })).toBeNull();
-    }
-    expect(
-      screen.queryByRole("textbox", { name: "Custom inactivity days" }),
-    ).toBeNull();
-    screen.getByRole("button", { name: "Configure automatic cleanup" });
-  });
-
-  it("mounts the threshold editor on Configure and unmounts it on Collapse", async () => {
-    renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({ enabled: true }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    await openThresholdEditor();
-    // The fuller safety explanation belongs to the decision, so it arrives
-    // with the editor rather than sitting on the collapsed row.
-    screen.getByText(/Age alone never makes one safe/);
-    screen.getByRole("button", { name: "30 days" });
-
-    fireEvent.click(
-      screen.getByRole("button", {
-        name: "Collapse automatic cleanup settings",
-      }),
-    );
-    await waitFor(() => {
-      expect(
-        screen.queryByRole("textbox", { name: "Custom inactivity days" }),
-      ).toBeNull();
-    });
-    expect(screen.queryByRole("button", { name: "30 days" })).toBeNull();
-    screen.getByRole("button", { name: "Configure automatic cleanup" });
-  });
-
-  it("toggles the policy from the collapsed summary without unfolding it", async () => {
-    const requests: boolean[] = [];
-    let policy = policyFixture({ enabled: true, revision: 4 });
-    renderSection(
-      clientWithPolicy({
-        get: () => policy,
-        set: (request) => {
-          requests.push(request.enabled);
-          policy = policyFixture({
-            ...request,
-            revision: request.expectedRevision + 1,
-          });
-          return policy;
-        },
-      }),
-    );
-
-    fireEvent.click(await loadedCleanupToggle());
-
-    await waitFor(() => {
-      expect(requests).toEqual([false]);
-    });
-    // Off takes the disclosure with it - and the summary states the new state
-    // rather than leaving the threshold on screen.
-    await waitFor(() => {
-      expect(
-        screen.getByTestId("worktree-auto-cleanup-summary").textContent,
-      ).toBe("Automatic cleanup · Off");
-    });
-    expect(screen.queryByRole("button", { name: /configure/i })).toBeNull();
-  });
-
-  it("returns to the collapsed summary on a remount", async () => {
-    const client = clientWithPolicy({
-      get: () => policyFixture({ enabled: true }),
-      set: (r) => policyFixture(r),
-    });
-    const { unmount } = renderSection(client);
-
-    await openThresholdEditor();
-    unmount();
-
-    // Disclosure state is component-local and never persisted, so re-entering
-    // Settings - or switching hosts - starts from the summary again.
-    renderSection(client);
-    await screen.findByRole("button", { name: "Configure automatic cleanup" });
-    expect(
-      screen.queryByRole("textbox", { name: "Custom inactivity days" }),
-    ).toBeNull();
-  });
-
-  it("starts collapsed again when the scoped host changes in place", async () => {
-    const client = clientWithPolicy({
-      get: () => policyFixture({ enabled: true }),
-      set: (r) => policyFixture(r),
-    });
-    const { rerender } = renderSection(client);
-    await openThresholdEditor();
-
-    // The sidebar can switch straight from one usable host to another, which
-    // keeps the section at the same tree position and only swaps its scope.
-    // An open disclosure must not follow the panel across that switch: the
-    // new host starts from the summary like any first visit.
-    rerender(sectionForHost(client, HOST_B));
-
-    await screen.findByRole("button", { name: "Configure automatic cleanup" });
-    expect(
-      screen.queryByRole("textbox", { name: "Custom inactivity days" }),
-    ).toBeNull();
-  });
-
-  it("brings itself into view for a deep link naming its host, then drops the request", async () => {
-    // The Sweep dialog's discovery line leaves this one-shot request behind.
-    // The card is what the link promised, so it scrolls into view AND takes
-    // the caret - and clears the request, or a later visit to Settings would
-    // re-scroll to a card nobody asked about that time.
-    const scrolls = spyOnScrollIntoView();
-    useWorktreeCleanupViewStore.getState().requestAutoCleanupFocus("host-a");
-
-    renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({}),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    const row = screen.getByTestId("worktree-auto-cleanup-summary-row");
-    await waitFor(() => {
-      expect(scrolls.mock.contexts).toContain(row);
-    });
-    expect(document.activeElement).toBe(row);
-    expect(
-      useWorktreeCleanupViewStore.getState().autoCleanupFocusHostId,
-    ).toBeNull();
-  });
-
-  it("leaves another host's focus request alone until that host is scoped", async () => {
-    // The request outlives its destination whenever the named host never
-    // mounts a card - offline, too old, or Settings simply never opened. An
-    // unscoped flag would then be spent on whichever host came next, scrolling
-    // to a machine nobody asked about.
-    const scrolls = spyOnScrollIntoView();
-    useWorktreeCleanupViewStore.getState().requestAutoCleanupFocus("host-b");
-    const client = clientWithPolicy({
-      get: () => policyFixture({}),
-      set: (r) => policyFixture(r),
-    });
-
-    const { rerender } = renderSection(client);
-
-    await loadedCleanupToggle();
-    expect(scrolls.mock.contexts).toEqual([]);
-    expect(document.activeElement).not.toBe(
-      screen.getByTestId("worktree-auto-cleanup-summary-row"),
-    );
-    // Untouched, so it is still there for the host it was actually about.
-    expect(useWorktreeCleanupViewStore.getState().autoCleanupFocusHostId).toBe(
-      "host-b",
-    );
-
-    rerender(sectionForHost(client, HOST_B));
-
-    const row = screen.getByTestId("worktree-auto-cleanup-summary-row");
-    await waitFor(() => {
-      expect(scrolls.mock.contexts).toContain(row);
-    });
-    expect(
-      useWorktreeCleanupViewStore.getState().autoCleanupFocusHostId,
-    ).toBeNull();
-  });
-
-  it("says the host is too old rather than offering a client-side fallback", () => {
-    state.supported = false;
-
-    renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({}),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    const notice = screen.getByTestId("worktree-auto-cleanup-notice");
-    expect(notice.getAttribute("data-gate")).toBe("unsupported");
-    expect(notice.textContent).toContain("Update the host to turn it on");
-    // No controls at all: the affordance is withheld, not disabled.
-    expect(screen.queryByRole("switch")).toBeNull();
-  });
-
-  it("explains an offline host without claiming its version is the problem", () => {
-    state.reachability = { status: "unreachable", hostLabel: "Host A" };
-
-    renderSection(null);
-
-    const notice = screen.getByTestId("worktree-auto-cleanup-notice");
-    expect(notice.getAttribute("data-gate")).toBe("offline");
-    expect(notice.textContent).toContain("Host A is offline");
-    expect(notice.textContent).not.toContain("Update the host");
-    expect(screen.queryByRole("switch")).toBeNull();
-  });
-
-  it("explains a pause in plain English and offers no repair affordance", async () => {
-    renderSection(
-      clientWithPolicy({
-        get: () =>
-          policyFixture({
-            enabled: true,
-            pausedReason: "needs_reauth",
-            lastEvaluatedAt: 1_000,
-            nextEvaluationAt: null,
-          }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    await waitFor(() => {
-      screen.getByTestId("worktree-auto-cleanup-paused");
-    });
-    const paused = screen.getByTestId("worktree-auto-cleanup-paused");
-    expect(paused.textContent).toContain("needs to be re-authorized");
-    expect(paused.textContent).toContain("resumes on its own");
-    // A `null` next check while paused is a real state, not an unknown one.
-    expect(
-      screen.getByTestId("worktree-auto-cleanup-schedule").textContent,
-    ).toContain("Next check: paused");
-    // Nothing to press: every pause arm clears without the user acting.
-    expect(screen.queryByRole("button", { name: /re-?authorize/i })).toBeNull();
-  });
-
-  it("shows the history button only while automatic cleanup is enabled", async () => {
-    // The history is automatic runs only - manual deletions never appear in
-    // it - so with the policy off the button is hidden even when earlier runs
-    // exist. A button beside "Cleanup is off" read as the place manual
-    // deletions should show up.
-    const { unmount } = renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({ enabled: false, lastEvaluatedAt: 1_000 }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-    await waitFor(() => {
-      screen.getByRole("switch", { name: "Automatic cleanup" });
-    });
-    expect(
-      screen.queryByRole("button", { name: /automatic cleanup history/i }),
-    ).toBeNull();
-    unmount();
-
-    renderSection(
-      clientWithPolicy({
-        get: () => policyFixture({ enabled: true }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-    await waitFor(() => {
-      screen.getByRole("button", { name: /automatic cleanup history/i });
-    });
-  });
-
-  it("renders an upcoming check as a countdown, never as a past-tense label", async () => {
-    renderSection(
-      clientWithPolicy({
-        get: () =>
-          policyFixture({
-            enabled: true,
-            lastEvaluatedAt: Date.now() - 5 * 60_000,
-            nextEvaluationAt: Date.now() + 10 * 60_000,
-          }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    await waitFor(() => {
-      screen.getByTestId("worktree-auto-cleanup-schedule");
-    });
-    const schedule = screen.getByTestId("worktree-auto-cleanup-schedule");
-    // The regression this pins: `useRelativeTimestamp` clamps a negative
-    // delta, so a check ~10m AWAY rendered as "Just now" - a past-tense claim
-    // about an event that has not happened.
-    expect(schedule.textContent).toMatch(/next check in \d+m/);
-    expect(schedule.textContent).not.toContain("next check Just now");
-    expect(schedule.textContent).toContain("Last checked");
-  });
-
-  it("renders a sub-minute check as under a minute rather than a seconds count", async () => {
-    // The shared clock ticks once a minute, so a seconds count would sit
-    // frozen past its own deadline. The phrase stays true for the whole tick.
-    renderSection(
-      clientWithPolicy({
-        get: () =>
-          policyFixture({
-            enabled: true,
-            lastEvaluatedAt: Date.now() - 5 * 60_000,
-            nextEvaluationAt: Date.now() + 30_000,
-          }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    await waitFor(() => {
-      screen.getByTestId("worktree-auto-cleanup-schedule");
-    });
-    expect(
-      screen.getByTestId("worktree-auto-cleanup-schedule").textContent,
-    ).toContain("next check in under a minute");
-  });
-
-  it("flips to due now at the deadline without waiting for the next clock tick", async () => {
-    // The shared clock samples once a minute; a deadline landing between two
-    // samples must not leave "in under a minute" on screen past itself.
-    renderSection(
-      clientWithPolicy({
-        get: () =>
-          policyFixture({
-            enabled: true,
-            lastEvaluatedAt: Date.now() - 5 * 60_000,
-            nextEvaluationAt: Date.now() + 700,
-          }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    await waitFor(() => {
-      expect(
-        screen.getByTestId("worktree-auto-cleanup-schedule").textContent,
-      ).toContain("next check in under a minute");
-    });
-    await waitFor(
-      () => {
-        expect(
-          screen.getByTestId("worktree-auto-cleanup-schedule").textContent,
-        ).toContain("next check due now");
-      },
-      { timeout: 3_000 },
-    );
-  });
-
-  it("renders an overdue check as due now rather than counting down to zero", async () => {
-    renderSection(
-      clientWithPolicy({
-        get: () =>
-          policyFixture({
-            enabled: true,
-            lastEvaluatedAt: Date.now() - 60 * 60_000,
-            nextEvaluationAt: Date.now() - 60_000,
-          }),
-        set: (r) => policyFixture(r),
-      }),
-    );
-
-    await waitFor(() => {
-      screen.getByTestId("worktree-auto-cleanup-schedule");
-    });
-    expect(
-      screen.getByTestId("worktree-auto-cleanup-schedule").textContent,
-    ).toContain("next check due now");
-  });
-
-  it("sends the current revision with a preset threshold change", async () => {
-    const requests: Array<{
-      readonly inactivityDays: number;
-      readonly expectedRevision: number;
-    }> = [];
-    let policy = policyFixture({ enabled: true, revision: 3 });
-    renderSection(
-      clientWithPolicy({
-        get: () => policy,
-        set: (request) => {
-          requests.push(request);
-          policy = policyFixture({
-            ...request,
-            revision: request.expectedRevision + 1,
-          });
-          return policy;
-        },
-      }),
-    );
-
-    await openThresholdEditor();
-    fireEvent.click(screen.getByRole("button", { name: "7 days" }));
-
-    await waitFor(() => {
-      expect(requests).toEqual([
-        { enabled: true, inactivityDays: 7, expectedRevision: 3 },
-      ]);
-    });
-    // The response is the fresh state, so the control re-presents it without a
-    // second round trip.
-    await waitFor(() => {
-      expect(
-        screen
-          .getByRole("button", { name: "7 days" })
-          .getAttribute("aria-pressed"),
-      ).toBe("true");
-    });
-  });
-
-  it("offers only the presets inside the host's bounds", async () => {
-    const requests: number[] = [];
-    renderSection(
-      clientWithPolicy({
-        get: () =>
-          policyFixture({
-            enabled: true,
-            inactivityDays: 30,
-            bounds: { minDays: 30, maxDays: 60 },
-          }),
-        set: (request) => {
-          requests.push(request.inactivityDays);
-          return policyFixture(request);
-        },
-      }),
-    );
-
-    await openThresholdEditor();
-    // Only the two presets inside [30, 60] are offered - a preset the host
-    // would refuse is not shown as a button at all.
-    screen.getByRole("button", { name: "30 days" });
-    screen.getByRole("button", { name: "60 days" });
-    for (const days of [7, 14, 90]) {
-      expect(screen.queryByRole("button", { name: `${days} days` })).toBeNull();
-    }
-
-    // The custom input validates against the same bounds.
-    const input = screen.getByRole("textbox", {
-      name: "Custom inactivity days",
-    });
-    fireEvent.change(input, { target: { value: "90" } });
-    fireEvent.blur(input);
-    await waitFor(() => {
-      screen.getByText("Choose between 30 and 60 days.");
-    });
-    expect(requests).toEqual([]);
-
-    fireEvent.change(input, { target: { value: "45" } });
-    fireEvent.blur(input);
-    await waitFor(() => {
-      expect(requests).toEqual([45]);
-    });
-  });
-
-  it("refuses a custom value outside the host's bounds without sending it", async () => {
-    const requests: number[] = [];
-    renderSection(
-      clientWithPolicy({
-        get: () =>
-          policyFixture({
-            enabled: true,
-            bounds: { minDays: 1, maxDays: 365 },
-          }),
-        set: (request) => {
-          requests.push(request.inactivityDays);
-          return policyFixture(request);
-        },
-      }),
-    );
-
-    await openThresholdEditor();
-    const input = screen.getByRole("textbox", {
-      name: "Custom inactivity days",
-    });
-    fireEvent.change(input, { target: { value: "400" } });
-    fireEvent.blur(input);
-
-    await waitFor(() => {
-      screen.getByText("Choose between 1 and 365 days.");
-    });
-    expect(requests).toEqual([]);
-
-    fireEvent.change(input, { target: { value: "45" } });
-    fireEvent.blur(input);
-    await waitFor(() => {
-      expect(requests).toEqual([45]);
-    });
-  });
-
-  it("re-reads and explains a revision conflict instead of retrying blind", async () => {
-    let policy = policyFixture({ enabled: false, revision: 1 });
-    let setCalls = 0;
-    renderSection(
-      clientWithPolicy({
-        get: () => policy,
-        set: () => {
-          setCalls += 1;
-          // Another surface moved the policy on underneath this client.
-          policy = policyFixture({
-            enabled: true,
-            inactivityDays: 14,
-            revision: 2,
-          });
-          throw new HostRpcError({
-            code: "AUTO_CLEANUP_POLICY_REVISION_CONFLICT",
-            message: "Policy revision conflict",
-            requestId: "req-conflict",
-            method: "worktree.setAutoCleanupPolicy",
-            fatalDetails: null,
-          });
-        },
-      }),
-    );
-
-    fireEvent.click(await loadedCleanupToggle());
-
-    await waitFor(() => {
-      screen.getByTestId("worktree-auto-cleanup-conflict");
-    });
-    expect(setCalls).toBe(1);
-    // The re-read wins: the summary states what actually landed, and the
-    // editor behind Configure agrees with it.
-    await waitFor(() => {
-      expect(
-        screen.getByTestId("worktree-auto-cleanup-summary").textContent,
-      ).toBe("Automatic cleanup · On · after 14 days");
-    });
-    await openThresholdEditor();
-    expect(
-      screen
-        .getByRole("button", { name: "14 days" })
-        .getAttribute("aria-pressed"),
-    ).toBe("true");
-  });
-});
