+++ clients/gui-app/src/components/settings/panels/__tests__/worktrees-row-overflow.test.tsx
@@ -80,6 +80,7 @@ function renderSingleRow(
         onVisiblePathsChange={() => {}}
         taskTitlesByEpicId={new Map()}
         toolbarProps={{
+          cleanup: null,
           onRefresh: () => Promise.resolve(),
           refreshing: false,
           canRefresh: true,
