+++ clients/gui-app/src/lib/worktree/open-auto-cleanup-settings.ts
@@ -16,15 +16,16 @@ import { carryViewedHostIntoSettingsScope } from "@/components/settings/host-sco
  * the same order and for the same reasons: the host scope and the sub-view are
  * applied BEFORE navigating, so the panel reads them on its first render rather
  * than flashing the wrong host or the cleanup history. The difference is the
- * third hint - a notification names a run, this names the card - and it is a
- * hint in exactly the same sense: a host that is offline or too old renders no
- * card at all, and the destination still has to be the worktrees panel.
+ * third hint - a notification names a run, this names the cleanup policy - and
+ * it is a hint in exactly the same sense: a host that is offline or too old
+ * opens no policy popover at all, and the destination still has to be the
+ * worktrees panel.
  *
  * `hostId` is the CALLING surface's latched host. `null` leaves Settings
  * administering whatever host it already was, which is the honest answer when
- * the caller has no machine of its own to name - and it asks for no card
- * focus either, because a request that cannot name its host is a request the
- * card cannot check before acting on.
+ * the caller has no machine of its own to name - and it asks for no popover
+ * either, because a request that cannot name its host is a request the chip
+ * cannot check before acting on.
  */
 export function openWorktreeAutoCleanupSettings(
   navigate: UseNavigateResult<string>,
